feat: add document state detection
Added a new feature to detect the document state based on the estimated distance between the camera and the document.

The state can be:
- `ComeCloser`: When the document is too far away.
- `Correct`: When the document is at the ideal distance.
- `GoFurther`: When the document is too close.

This state is included in the `ScannedDocumentResult` and can be used to provide feedback to the user during the scanning process.

(cherry picked from commit 6324e9d92bb8b4695ef5a84139f2360c722f8f3e)