Upgrade Kotlin to 2.1.0 and Gradle to 8.7.3
Upgrade Kotlin to version 2.1.0, Gradle to 8.7.3, and enable when guards for the Kotlin compiler.
Update the CameraView composable to accept a nullable Analyzer and avoid binding it if it is null.
Modify the GraphicOverlay to include an onPointsUpdate callback and call it after transforming the points.

(cherry picked from commit e3866ed63e8f4356704946564d0342996f6ac758)