publish aar with sources and javadoc
This commit updates the build configuration to include sources and Javadoc JARs when publishing the AAR. It sets the base name for the AAR and uses the 'release' variant for publishing. It also includes tasks to generate sources and Javadoc JARs, including a task for Dokka if used. Additionally, it modifies the output file name for the AAR artifact.